pyproject.toml
[tool.poetry]
name = "repl-nix-careers-website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_careers_website"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
app = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
